Update level 1 morningstar

+++ game/Camera.cpp
@@ -25,7 +25,7 @@ void Camera::Update()
 		_viewport.x = MapWidth - Window_Width; 
 
 }
-D3DXVECTOR2 Camera::Transform(int x, int y)
+D3DXVECTOR2 Camera::Transform(float x, float y)
 {
 	return D3DXVECTOR2(x - _viewport.x, y - _viewport.y);
 }
